requirements.txt
discord.py
youtube_dl
PyNaCl
ffmpeg-python
python-ffmpeg
ffmpeg
matplotlib
tensorflow
keras
xlrd